perf(spec-parser): refactor code for browser environment
