+++ packages/fx-core/src/common/spec-parser/manifestUpdater.ts
@@ -3,18 +3,18 @@
 "use strict";
 
 import { OpenAPIV3 } from "openapi-types";
-import { Parameter, ErrorType, WarningResult, WarningType } from "./interfaces";
 import fs from "fs-extra";
 import path from "path";
-import { getRelativePath, updateFirstLetter } from "./utils";
+import { ErrorType, WarningResult } from "./interfaces";
+import { getRelativePath, parseApiInfo } from "./utils";
 import { SpecParserError } from "./specParserError";
 import { ConstantString } from "./constants";
-import { format } from "util";
 import {
   IComposeExtension,
   IMessagingExtensionCommand,
   TeamsAppManifest,
 } from "@microsoft/teamsfx-api";
+
 export async function updateManifest(
   manifestPath: string,
   outputSpecPath: string,
@@ -22,7 +22,6 @@ export async function updateManifest(
   spec: OpenAPIV3.Document
 ): Promise<[TeamsAppManifest, WarningResult[]]> {
   try {
-    // TODO: manifest interface can be updated when manifest parser library is ready
     const originalManifest: TeamsAppManifest = await fs.readJSON(manifestPath);
 
     const [commands, warnings] = await generateCommands(spec, adaptiveCardFolder, manifestPath);
@@ -48,51 +47,6 @@ export async function updateManifest(
   }
 }
 
-export function generateParametersFromSchema(
-  schema: OpenAPIV3.SchemaObject,
-  name: string,
-  isRequired = false
-): [Parameter[], Parameter[]] {
-  const requiredParams: Parameter[] = [];
-  const optionalParams: Parameter[] = [];
-
-  if (
-    schema.type === "string" ||
-    schema.type === "integer" ||
-    schema.type === "boolean" ||
-    schema.type === "number"
-  ) {
-    const parameter = {
-      name: name,
-      title: updateFirstLetter(name),
-      description: schema.description ?? "",
-    };
-    if (isRequired && schema.default === undefined) {
-      requiredParams.push(parameter);
-    } else {
-      optionalParams.push(parameter);
-    }
-  } else if (schema.type === "object") {
-    const { properties } = schema;
-    for (const property in properties) {
-      let isRequired = false;
-      if (schema.required && schema.required?.indexOf(property) >= 0) {
-        isRequired = true;
-      }
-      const [requiredP, optionalP] = generateParametersFromSchema(
-        properties[property] as OpenAPIV3.SchemaObject,
-        property,
-        isRequired
-      );
-
-      requiredParams.push(...requiredP);
-      optionalParams.push(...optionalP);
-    }
-  }
-
-  return [requiredParams, optionalParams];
-}
-
 export async function generateCommands(
   spec: OpenAPIV3.Document,
   adaptiveCardFolder: string,
@@ -112,76 +66,18 @@ export async function generateCommands(
           if (method === ConstantString.PostMethod || method === ConstantString.GetMethod) {
             const operationItem = operations[method];
             if (operationItem) {
-              const requiredParams: Parameter[] = [];
-              const optionalParams: Parameter[] = [];
-              const paramObject = operationItem.parameters as OpenAPIV3.ParameterObject[];
+              const [command, warning] = parseApiInfo(operationItem);
 
-              if (paramObject) {
-                paramObject.forEach((param: OpenAPIV3.ParameterObject) => {
-                  const parameter: Parameter = {
-                    name: param.name,
-                    title: updateFirstLetter(param.name),
-                    description: param.description ?? "",
-                  };
+              const adaptiveCardPath = path.join(adaptiveCardFolder, command.id + ".json");
+              command.apiResponseRenderingTemplateFile = (await fs.pathExists(adaptiveCardPath))
+                ? getRelativePath(manifestPath, adaptiveCardPath)
+                : "";
 
-                  const schema = param.schema as OpenAPIV3.SchemaObject;
-                  if (param.in !== "header" && param.in !== "cookie") {
-                    if (param.required && schema?.default === undefined) {
-                      requiredParams.push(parameter);
-                    } else {
-                      optionalParams.push(parameter);
-                    }
-                  }
-                });
+              if (warning) {
+                warnings.push(warning);
               }
 
-              if (operationItem.requestBody) {
-                const requestBody = operationItem.requestBody as OpenAPIV3.RequestBodyObject;
-                const requestJson = requestBody.content["application/json"];
-                if (Object.keys(requestJson).length !== 0) {
-                  const schema = requestJson.schema as OpenAPIV3.SchemaObject;
-                  const [requiredP, optionalP] = generateParametersFromSchema(
-                    schema,
-                    "requestBody",
-                    requestBody.required
-                  );
-                  requiredParams.push(...requiredP);
-                  optionalParams.push(...optionalP);
-                }
-              }
-
-              const operationId = operationItem.operationId!;
-
-              const adaptiveCardPath = path.join(adaptiveCardFolder, operationId + ".json");
-
-              const parameters = [];
-
-              if (requiredParams.length != 0) {
-                parameters.push(...requiredParams);
-              } else {
-                parameters.push(optionalParams[0]);
-              }
-
-              const command: IMessagingExtensionCommand = {
-                context: ["compose"],
-                type: "query",
-                title: operationItem.summary ?? "",
-                id: operationId,
-                parameters: parameters,
-                description: operationItem.description ?? "",
-                apiResponseRenderingTemplateFile: (await fs.pathExists(adaptiveCardPath))
-                  ? getRelativePath(manifestPath, adaptiveCardPath)
-                  : "",
-              };
               commands.push(command);
-
-              if (requiredParams.length === 0 && optionalParams.length > 1) {
-                warnings.push({
-                  type: WarningType.OperationOnlyContainsOptionalParam,
-                  content: format(ConstantString.OperationOnlyContainsOptionalParam, operationId),
-                  data: operationId,
-                });
-              }
             }
           }
         }
