+++ packages/fx-core/src/core/FxCore.ts
@@ -38,8 +38,7 @@ import { getLocalizedString } from "../common/localizeUtils";
 import { LaunchHelper } from "../common/m365/launchHelper";
 import { ListCollaboratorResult, PermissionsResult } from "../common/permissionInterface";
 import { isValidProjectV2, isValidProjectV3 } from "../common/projectSettingsHelper";
-import { SpecParser } from "../common/spec-parser/specParser";
-import { SpecParserError } from "../common/spec-parser/specParserError";
+import { SpecParser, SpecParserError } from "../common/spec-parser";
 import { VersionSource, VersionState } from "../common/versionMetadata";
 import { ILifecycle, LifecycleName } from "../component/configManager/interface";
 import { YamlParser } from "../component/configManager/parser";
