+++ packages/fx-core/tests/question/create.test.ts
@@ -22,8 +22,7 @@ import * as path from "path";
 import sinon from "sinon";
 import { FeatureFlagName } from "../../src/common/constants";
 import { getLocalizedString } from "../../src/common/localizeUtils";
-import { ErrorType, ValidationStatus, WarningType } from "../../src/common/spec-parser/interfaces";
-import { SpecParser } from "../../src/common/spec-parser/specParser";
+import { ErrorType, ValidationStatus, WarningType, SpecParser } from "../../src/common/spec-parser";
 import { AppDefinition } from "../../src/component/driver/teamsApp/interfaces/appdefinitions/appDefinition";
 import { manifestUtils } from "../../src/component/driver/teamsApp/utils/ManifestUtils";
 import { setTools } from "../../src/core/globalVars";
